Share the demo network and market modules.
