+++ demos/extension/src/config.ts
@@ -3,17 +3,13 @@ const RP_ORIGIN = "https://mera.category.xyz";
 const PASSKEY_PAGE_URL = `${RP_ORIGIN}/demo/passkey.html`;
 const PASSKEY_CHANNEL = "mera-passkey";
 const DEMO_CHAIN_ID = 31337;
-const RPC_URL = "https://evm-network-production.up.railway.app";
-const MAX_UINT256 = 2n ** 256n - 1n;
 const RECOVERY_VISIBLE_MS = 60_000;
 
 export {
   DEMO_CHAIN_ID,
-  MAX_UINT256,
   PASSKEY_CHANNEL,
   PASSKEY_PAGE_URL,
   RECOVERY_VISIBLE_MS,
   RP_ID,
   RP_ORIGIN,
-  RPC_URL,
 };
